Zadanie_3/requirements.txt
langgraph>=0.2.60
langchain>=0.3.10
langchain-ollama>=0.2.0
chromadb>=0.5.5
psycopg[binary]>=3.2.1
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.9.2
tavily-python>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
